Move remaining components into components dir
